+++ apps/web/src/utils/usernames.ts
@@ -8,6 +8,7 @@ import {
   labelhash,
   createPublicClient,
   http,
+  sha256,
 } from 'viem';
 import { normalize } from 'viem/ens';
 import RegistrarControllerABI from 'apps/web/src/abis/RegistrarControllerABI';
@@ -36,6 +37,26 @@ import {
 import { getBasenamePublicClient } from 'apps/web/src/hooks/useBasenameChain';
 import { USERNAME_L2_RESOLVER_ADDRESSES } from 'apps/web/src/addresses/usernames';
 import { logger } from 'apps/web/src/utils/logger';
+
+// Note: The animations provided by the studio team didn't match the number from our SVGs
+//       If we replace those, double check the animation avatar is the same shape as the SVG
+import animation1 from 'apps/web/src/components/Basenames/BasenameAvatar/animations/01.json';
+import animation2 from 'apps/web/src/components/Basenames/BasenameAvatar/animations/02.json';
+import animation3 from 'apps/web/src/components/Basenames/BasenameAvatar/animations/03.json';
+import animation4 from 'apps/web/src/components/Basenames/BasenameAvatar/animations/04.json';
+import animation5 from 'apps/web/src/components/Basenames/BasenameAvatar/animations/05.json';
+import animation6 from 'apps/web/src/components/Basenames/BasenameAvatar/animations/06.json';
+import animation7 from 'apps/web/src/components/Basenames/BasenameAvatar/animations/07.json';
+
+import image1 from 'apps/web/src/components/Basenames/BasenameAvatar/images/1.svg';
+import image2 from 'apps/web/src/components/Basenames/BasenameAvatar/images/2.svg';
+import image3 from 'apps/web/src/components/Basenames/BasenameAvatar/images/3.svg';
+import image4 from 'apps/web/src/components/Basenames/BasenameAvatar/images/4.svg';
+import image5 from 'apps/web/src/components/Basenames/BasenameAvatar/images/5.svg';
+import image6 from 'apps/web/src/components/Basenames/BasenameAvatar/images/6.svg';
+import image7 from 'apps/web/src/components/Basenames/BasenameAvatar/images/7.svg';
+
+import { StaticImageData } from 'next/image';
 
 export const USERNAME_MIN_CHARACTER_LENGTH = 3;
 export const USERNAME_MAX_CHARACTER_LENGTH = 20;
@@ -641,6 +662,41 @@ export function buildBasenameReclaimContract(
     functionName: 'reclaim',
   };
 }
+
+/* 
+  Basename avatar / animations
+*/
+
+export const getUsernamePictureIndex = (name: string, totalOptions: number) => {
+  const nameAsUint8Array = Uint8Array.from(name.split('').map((letter) => letter.charCodeAt(0)));
+  const hash = sha256(nameAsUint8Array);
+  const hashValue = parseInt(hash, 16);
+  const remainder = hashValue % totalOptions;
+  const selectedOption = remainder;
+  return selectedOption;
+};
+
+export const getBasenameAnimation = (username: string) => {
+  const animations = [
+    animation1,
+    animation2,
+    animation3,
+    animation4,
+    animation5,
+    animation6,
+    animation7,
+  ];
+  const profilePictureIndex = getUsernamePictureIndex(username, animations.length);
+  const selectedAnimation = animations[profilePictureIndex];
+  return selectedAnimation;
+};
+
+export const getBasenameImage = (username: string) => {
+  const images = [image1, image2, image3, image4, image5, image6, image7];
+  const profilePictureIndex = getUsernamePictureIndex(username, images.length);
+  const selectedAnimation = images[profilePictureIndex] as StaticImageData;
+  return selectedAnimation;
+};
 
 /*
   Feature flags
